requirements.txt
streamlit
sentence-transformers
faiss-cpu==1.13.2
numpy
PyPDF2
scikit-learn
rank-bm25
requests